Extracts out paths-related functions from utils
