+++ cargo-pgx/src/command/init.rs
@@ -12,8 +12,9 @@ use crate::command::version::pgx_default;
 use crate::CommandExecute;
 use eyre::{eyre, WrapErr};
 use owo_colors::OwoColorize;
+use pgx_paths::prefix_path;
 use pgx_pg_config::{PgConfig, PgConfigSelector, Pgx};
-use pgx_utils::{prefix_path, SUPPORTED_MAJOR_VERSIONS};
+use pgx_utils::SUPPORTED_MAJOR_VERSIONS;
 use rayon::prelude::*;
 
 use std::collections::HashMap;
