+++ pgx-paths/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "pgx-paths"
+edition = "2021"
+version = "0.5.0-beta.0"
+authors = ["ZomboDB, LLC <zombodb@gmail.com>"]
+license = "MIT"
+description = "A crate containing path and directory helpers for 'pgx'"
+homepage = "https://github.com/zombodb/pgx"
+repository = "https://github.com/zombodb/pgx"
+#documentation = "https://docs.rs/pgx-utils"
+readme = "README.md"
+rust-version = "1.58"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+eyre = "0.6.8"
+serde = { version = "1.0.137", features = [ "derive" ] }
+serde_derive = "1.0.137"
+serde_json = "1.0.81"
