+++ pgx-tests/Cargo.toml
@@ -36,6 +36,7 @@ owo-colors = "3.4.0"
 once_cell = "1.10.0"
 libc = "0.2.126"
 pgx = { path = "../pgx", default-features = false, version= "=0.5.0-beta.0" }
+pgx-paths = { path = "../pgx-paths", version= "=0.5.0-beta.0" }
 pgx-macros = { path = "../pgx-macros", version= "=0.5.0-beta.0" }
 pgx-pg-config = { path = "../pgx-pg-config", version= "=0.5.0-beta.0" }
 pgx-utils = { path = "../pgx-utils", version= "=0.5.0-beta.0" }
